# Conflicts:
#	pet_shop/src/main/java/com/example/s13firstspring/controllers/AnimalController.java
#	pet_shop/src/main/java/com/example/s13firstspring/controllers/BrandController.java
#	pet_shop/src/main/java/com/example/s13firstspring/models/repositories/AnimalRepository.java
#	pet_shop/src/main/java/com/example/s13firstspring/models/repositories/BrandRepository.java
#	pet_shop/src/main/java/com/example/s13firstspring/models/repositories/ProductRepository.java
#	pet_shop/src/main/java/com/example/s13firstspring/services/AnimalService.java
#	pet_shop/src/main/java/com/example/s13firstspring/services/BrandService.java

+++ pet_shop/src/main/java/com/example/s13firstspring/models/repositories/AnimalRepository.java
@@ -5,8 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnimalRepository extends JpaRepository<Animal,Long> {
+public interface AnimalRepository extends JpaRepository<Animal,Integer> {
 
     Animal findByName(String name);
-
 }
