+++ pet_shop/src/main/java/com/example/s13firstspring/models/repositories/DiscountRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DiscountRepository extends JpaRepository<Discount, Long> {
+public interface DiscountRepository extends JpaRepository<Discount, Integer> {
 
     Discount findByName(String name);
 }
